requirements.txt
accelerate==0.29.2
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argcomplete==3.5.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.9.0
boto3==1.35.52
botocore==1.35.52
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.5.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.1
comm==0.2.2
contourpy==1.2.1
controlnet-aux==0.0.9
cssselect2==0.7.0
cssutils==2.11.1
cycler==0.12.1
dacite==1.7.0
datasets==2.18.0
debugpy==1.8.1
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.29.1
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.2
fiftyone==1.0.1
fiftyone-brain==0.17.0
fiftyone_db==1.1.7
filelock==3.13.4
flash-attn==2.5.8
Flask==3.1.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.0.3
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
gradio==3.47.1
gradio_client==0.6.0
graphql-core==3.2.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.27.0
humanize==4.11.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflate64==1.0.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.24
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
kornia==0.6.8
lazy_loader==0.4
lightning==2.3.0
lightning-utilities==0.11.2
lit==18.1.2
lpips==0.1.4
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mongoengine==0.24.2
more-itertools==10.3.0
motor==3.5.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.4
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python-headless==4.10.0.84
orjson==3.10.0
overrides==7.7.0
packaging==24.0
pandas==1.3.2
pandocfilters==1.5.1
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.24.1
portalocker==2.8.2
pprintpp==0.4.0
priority==2.0.0
proglog==0.1.10
progressbar2==4.4.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.22.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pybcj==1.0.2
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.6.4
pydantic_core==2.16.3
pydash==8.0.3
pydub==0.25.1
Pygments==2.17.2
pymongo==4.8.0
pyparsing==3.1.2
pyppmd==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.9
python-utils==3.8.2
pytorch-lightning==2.3.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0
pyzmq==25.1.2
pyzstd==0.16.2
rarfile==4.2
referencing==0.34.0
regex==2022.10.31
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.3
safetensors==0.4.2
scikit-image==0.20.0
scikit-learn==1.4.2
scipy==1.9.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==1.45.0
setproctitle==1.3.3
shapely==2.0.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.37.2
strawberry-graphql==0.247.0
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
taskgroup==0.0.0a4
tenacity==9.0.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.6.7
tinycss2==1.2.1
tokenizers==0.21.0
tomli==2.0.1
toolz==0.12.1
torchmetrics==1.4.0.post0
torchsummary==1.5.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.2
transformers==4.47.1
triton==2.3.1
trl==0.8.6
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tyro==0.8.3
tzdata==2024.1
tzlocal==5.2
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.29.0
visual-genome==1.1.1
voxel51-eta==0.13.0
wandb==0.16.6
wcwidth==0.2.13
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.14.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1


